Automatically clone missing repos
Note that when adding a webhook to a GitHub repo for the first time, the
paylod will not contain a "ref" member. The Hookshot library will raise an
error when the "ref" member is missing, and the target repo will be neither
cloned nor synced. Another webhook must be triggered by an update to the
`gh-pages` branch in order for the target site to be published.